Corrections of docstrings in kdt/Algorithms.py. TODO: review the comments below.
-- Algorithms.py (bfsTree) ----------------------------------

"'Breadth-first' in the sense that all the vertices reachable in step i are added 

to the tree before any of the newly-reachable vertices' reachable
 vertices are 

explored."

A bit confusing. Rewrite/remove?

-- Algorithms.py (pathsHop) ---------------------------------

"Each element of the vector,except those equal to -1, denotes the starting 

vertex whose path extended to the corresponding vertex. In case of multiple paths 

potentially extending to a single vertex, the chosen source is arbitrary."

Will the end-user understand such a description?

-- Algorithms.py (normalizeEdgeWeights) ---------------------

Meaning has changed:
  - was: each edge's weight becomes 1 / outdegree({node})
  - now: each edge's weight becomes {weight}/outdegree({node})
If it were as it was, then all the outgoing or incomming edges
would become equal.

-- Algorithms.py (pageRank) ---------------------------------

'iterations' argument has been published.