-- Mat.py, line 25 (__init__) ----------------------------------------------
1) Specifying rows after columns is rather a strange order. Traditionally, rows are mentioned before columns (e.g. in indexing operators).

2) It would be good to say a couple of words about these kdt.pyCombBLAS.Obj1 and kdt.pyCombBLAS.Obj2. If they are not to be published, their mentioning should be removed from docstrings.

-- Mat.py, line 443 (load) -------------------------------------------------

par_IO argument is to be described.

-- Mat.py, line 598 (addFilter)
   Mat.py, line 635 (delFilter) --------------------------------------------

There is a confusing discrepancy in terminology. While most of the time, values in a matrix are referred to simply as 'matrix elements', here we are talking about some graph vertices. The more it is confusing due to the filter's being likely to accept not 'a vertex' (which has no reflection in a matrix), but a matrix element, i.e. 'an edge' (a weight of an edge, if more precisely).

-- Mat.py, line 664 (materializeFilter)
   Mat.py, line 670 (deMaterializeFilter) ----------------------------------

These functions should be either documented or _-hidden.

-- Mat.py, line 756 (eWiseApply) -------------------------------------------

Arguments allowANulls, allowBNulls, and allowIntersect should be described.

-- Mat.py, line 815 (__getitem__) ------------------------------------------

1) I cannot comprehend the following description: 'a ":"-slice denoting all vertices'

2) 'a Vec containing a contiguous range of monotonically increasing integers' - strictly monotonically increasing? Are repetitions allowed?

-- Mat.py, line 1040 (SpGEMM) (and other mentions of 'semiring') -----------

The description of SecondMaxSemiringObj should be checked, and, if it is correct, it may be worthy to rename this variable (since I do not see in what way it is "Max").

-- Mat.py, line 1180 (max) -------------------------------------------------

Why do we need the init argument? I see it as a very confusing way to reinvent math.max(myFunc(...), minimax).

-- Mat.py, line 1360 (__div__) ---------------------------------------------

"NOTE:  for v0.1, the second argument may be only a scalar" - what about v.0.2?