Initial commit of SEJITS+KDT work for filters.
Most of the specialization machinery is contained in the specializer/ directory.  Changes necessary
to enable functionality are contained in a few KDT files.  In terms of functionality, the only regression
is that only a single filter is currently allowed (no filter stacks).  Other functionality is unimplemented
in the specializer, but should not affect non-SEJITS functionality.